Time: 0 ms (100.00%) | Memory: 7.1 MB (18.45%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.cpp
@@ -6,7 +6,7 @@ public:
             return 0;
         }
         int left=0;
-       long long right = static_cast<long long>(sqrt(c));
+       long long right=sqrt(c);
        
        
         while (left <= right) {
